Fix System Nexus lint errors
