+++ packages/workflow/src/nexus/system/payload-converter.ts
@@ -6,7 +6,8 @@ import {
   type TypeInfo,
   fromPayloadWithTypeInfo,
 } from '@temporalio/common';
-import { operationRegistry } from './generated/services';
+import { operationRegistry } from './generated/registry';
+import { withSystemNexusUserPayloadConverter } from './user-payload-converter';
 
 export function isSystemNexusOperation(service: string | undefined, operation: string | undefined): boolean {
   return (
@@ -16,42 +17,11 @@ export function isSystemNexusOperation(service: string | undefined, operation: s
   );
 }
 
-export function systemNexusOperationDefinition(service: string, operation: string) {
+export function systemNexusOperationDefinition(
+  service: string,
+  operation: string
+): (typeof operationRegistry)[number] | undefined {
   return operationRegistry.find((entry) => entry.service === service && entry.operation === operation);
-}
-
-let currentUserPayloadConverter: PayloadConverter | undefined;
-
-/**
- * Makes the current System Nexus operation's context-bound user converter
- * available to generated Workflow Service support conversion helpers.
- *
- * @internal
- */
-export function withSystemNexusUserPayloadConverter<T>(
-  converter: PayloadConverter,
-  context: SerializationContext | undefined,
-  fn: () => T
-): T {
-  const previous = currentUserPayloadConverter;
-  currentUserPayloadConverter = {
-    toPayload: (value, _context, hint) => converter.toPayload(value, context, hint),
-    fromPayload: (payload, _context, hint) => converter.fromPayload(payload, context, hint),
-    validateConverterHint: converter.validateConverterHint?.bind(converter),
-  };
-  try {
-    return fn();
-  } finally {
-    currentUserPayloadConverter = previous;
-  }
-}
-
-/** @internal Returns the user converter scoped by the System Nexus outer envelope. */
-export function currentSystemNexusUserPayloadConverter(): PayloadConverter {
-  if (currentUserPayloadConverter == null) {
-    throw new Error('System Nexus user payload converter context is not active');
-  }
-  return currentUserPayloadConverter;
 }
 
 /** Decode the JSON envelope returned by the Worker runtime to its generated protobuf shape. */
